requirements.txt
fastapi
uvicorn
pydantic
numpy
pillow
tensorflow
torch
scikit-learn
firebase-admin
google-generativeai
